requirements.txt
bidict==0.22.0
click==8.1.3
Flask==2.2.2
Flask-SocketIO==5.3.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymemcache==3.5.2
python-engineio==4.3.4
python-socketio==5.7.1
PyYAML==6.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
